NestedHashMap-based implementation of BQSRv2 along with a few minor optimizations.  Not a huge runtime upgrade over the long bitset approach, but it allows us to implement further optimizations going forward.  Integration test change because the original version had a bug in the quantized qual table creation.

+++ public/java/test/org/broadinstitute/sting/gatk/walkers/bqsr/CycleCovariateUnitTest.java
@@ -47,7 +47,7 @@ public class CycleCovariateUnitTest {
         verifyCovariateArray(readCovariates.getMismatchesKeySet(), -1, -1);
     }
 
-    private void verifyCovariateArray(long[][] values, int init, int increment) {
+    private void verifyCovariateArray(int[][] values, int init, int increment) {
         for (short i = 0; i < values.length; i++) {
             short actual = Short.decode(covariate.formatKey(values[i][0]));
             int expected = init + (increment * i);
